feat:용병 모집 API에 토큰 검증 추가

+++ src/main/java/com/shootdoori/match/service/MercenaryRecruitmentService.java
@@ -7,6 +7,7 @@ import com.shootdoori.match.entity.mercenary.MercenaryRecruitment;
 import com.shootdoori.match.entity.mercenary.MercenaryPosition;
 import com.shootdoori.match.entity.mercenary.MercenaryRecruitmentSkillLevel;
 import com.shootdoori.match.entity.team.Team;
+import com.shootdoori.match.exception.common.NoPermissionException;
 import com.shootdoori.match.exception.common.NotFoundException;
 import com.shootdoori.match.exception.common.ErrorCode;
 import com.shootdoori.match.repository.MercenaryRecruitmentRepository;
@@ -15,6 +16,8 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.util.Objects;
 
 @Service
 @Transactional
@@ -27,11 +30,14 @@ public class MercenaryRecruitmentService {
         this.teamRepository = teamRepository;
     }
 
-    public RecruitmentResponse create(RecruitmentCreateRequest request) {
-        // TODO: 요청을 보낸 사용자가 이 게시글을 생성할 권한이 있는지 확인하는 로직 추가
+    public RecruitmentResponse create(RecruitmentCreateRequest request, Long loginUserId) {
 
         Team team = teamRepository.findById(request.teamId()).orElseThrow(
             () -> new NotFoundException(ErrorCode.TEAM_NOT_FOUND, String.valueOf(request.teamId())));
+
+        if(!team.getCaptain().getId().equals(loginUserId)) {
+            throw new NoPermissionException();
+        }
 
         MercenaryPosition position = MercenaryPosition.fromDisplayName(request.position());
         MercenaryRecruitmentSkillLevel skillLevel = MercenaryRecruitmentSkillLevel.fromDisplayName(request.skillLevel());
@@ -57,11 +63,14 @@ public class MercenaryRecruitmentService {
         return new RecruitmentResponse(recruitment);
     }
 
-    public RecruitmentResponse update(Long id, RecruitmentUpdateRequest updateRequest) {
-        // TODO: 요청을 보낸 사용자가 이 게시글을 수정할 권한이 있는지 확인하는 로직 추가
+    public RecruitmentResponse update(Long id, RecruitmentUpdateRequest updateRequest, Long loginUserId) {
 
         MercenaryRecruitment recruitment = recruitmentRepository.findById(id)
             .orElseThrow(() -> new NotFoundException(ErrorCode.RECRUITMENT_NOT_FOUND));
+
+        if(!recruitment.getTeam().getCaptain().getId().equals(loginUserId)) {
+            throw new NoPermissionException();
+        }
 
         MercenaryPosition position = MercenaryPosition.fromDisplayName(updateRequest.position());
         MercenaryRecruitmentSkillLevel skillLevel = MercenaryRecruitmentSkillLevel.fromDisplayName(updateRequest.skillLevel());
@@ -72,11 +81,13 @@ public class MercenaryRecruitmentService {
         return new RecruitmentResponse(recruitment);
     }
 
-    public void delete(Long id) {
-        // TODO: 요청을 보낸 사용자가 이 게시글을 삭제할 권한이 있는지 확인하는 로직 추가
+    public void delete(Long id, Long loginUserId) {
 
-        recruitmentRepository.findById(id)
-            .orElseThrow(() -> new NotFoundException(ErrorCode.RECRUITMENT_NOT_FOUND));
+        MercenaryRecruitment recruitment = recruitmentRepository.findById(id).orElseThrow(() -> new NotFoundException(ErrorCode.RECRUITMENT_NOT_FOUND));
+
+        if(!recruitment.getTeam().getCaptain().getId().equals(loginUserId)) {
+            throw new NoPermissionException();
+        }
 
         recruitmentRepository.deleteById(id);
     }
